+++ test/javascripts/shopify_app/app_bridge_redirect_test.js
@@ -1,66 +1,21 @@
+const { assert } = require("chai");
+
 suite('appBridgeRedirect', () => {
   const sandbox = sinon.createSandbox();
   const url = '/settings';
-  const originalAppBridge = window['app-bridge'];
-
-  setup(() => {
-    window['app-bridge'] = {
-      default() {},
-      actions: {
-        Redirect: {
-          Action: {
-            REMOTE: 'REDIRECT::REMOTE',
-          },
-          create() {
-            return {
-              dispatch() {},
-            };
-          },
-        },
-      },
-    };
-    document.body.dataset.apiKey = '123';
-    document.body.dataset.shopOrigin = 'myshop.com';
-    document.body.dataset.host = 'https://mock-host/admin';
-  });
 
   teardown(() => {
     sandbox.restore();
-    window['app-bridge'] = originalAppBridge;
   });
 
-
-  test('calls App Bridge to create an app with the apiKey and shopOrigin from window', () => {
-    var createApp = sinon.spy();
-    sinon.stub(window['app-bridge'], 'default').callsFake(createApp);
+  test('calls App Bridge redirect to normalized url', () => {
+    const open = sinon.spy();
+    sandbox.stub(window, 'open').callsFake(open);
 
     appBridgeRedirect(url);
 
-    sinon.assert.calledWith(createApp, {
-      apiKey: '123',
-      host: 'https://mock-host/admin',
-    });
-  });
-
-  test('calls to dispatch a remote redirect App Bridge action with the url normalized to be relative to the window origin', () => {
-    const AppBridge = window['app-bridge'];
-    const Redirect = AppBridge.actions.Redirect;
-    var mockApp = {};
-    var RedirectInstance = {
-      dispatch: sinon.spy(),
-    };
-    sinon.stub(AppBridge, 'default').callsFake(() => mockApp);
-    sinon.stub(Redirect, 'create').callsFake(() => RedirectInstance);
-
-    const normalizedUrl = `${window.location.origin}${url}`;
-
-    appBridgeRedirect(url);
-
-    sinon.assert.calledWith(Redirect.create, mockApp);
-    sinon.assert.calledWith(
-        RedirectInstance.dispatch,
-        'REDIRECT::REMOTE',
-        normalizedUrl
-    );
+    assert(open.calledOnce);
+    assert.match(open.lastCall.args[0], new RegExp(`${url}$`));
+    assert.equal(open.lastCall.args[1], '_top');
   });
 });